requirements.txt
asgiref==3.8.1
Django==5.0.4
sqlparse==0.4.4
tzdata==2024.1
djangorestframework==3.14.0
channels==4.0.0
daphne==4.0.0
django-geojson==4.0.0
folium==0.14.0
django-notifications-hq==1.8.2
pandas==2.2.0
matplotlib==3.8.2
requests==2.31.0
geopy==2.4.1
websockets==12.0
